// packages/framework/metafox-core/src/types.tsx
export type BasicPlaceItem = {
  address?: string;
  name?: string;
  lat: number;
  lng: number;
};